cleaned up redundant pages, added favicon
