+++ src/components/Nav.js
@@ -11,6 +11,8 @@ class Nav extends Component {
     };
   }
 
+  // fetching all recipes here so there are no redundant calls throughout surprise and all recipes page
+  
   componentDidMount() {
     fetch(
       `http://localhost:3333/recipenames`,
